+++ test/codegen/language-mappers/zod-to-string-mapper.test.ts
@@ -152,6 +152,20 @@ describe('ZodToStringMapper', () => {
       expect(rendered).to.equal('z.object({name: z.string(); age: z.number()})')
     })
 
+    it('Can successfully parse records', () => {
+      const zodAst = secureEvaluateSchema(`z.record(z.string(), z.number())`)
+      const mapper = new ZodToStringMapper()
+      const rendered = mapper.renderField(zodAst.schema!)
+      expect(rendered).to.equal('z.record(z.string(), z.number())')
+    })
+
+    it('Can successfully parse records with object values', () => {
+      const zodAst = secureEvaluateSchema(`z.record(z.string(), z.object({ id: z.string(), count: z.number() }))`)
+      const mapper = new ZodToStringMapper()
+      const rendered = mapper.renderField(zodAst.schema!)
+      expect(rendered).to.equal('z.record(z.string(), z.object({id: z.string(); count: z.number()}))')
+    })
+
     it('Can successfully parse an optional wrapped type', () => {
       const zodAst = secureEvaluateSchema(`z.optional(z.string())`)