+++ test/codegen/language-mappers/zod-to-typescript-mapper.test.ts
@@ -152,6 +152,20 @@ describe('ZodToTypescriptMapper', () => {
       expect(rendered).to.equal('"someKey": { "name": string; "age": number }')
     })
 
+    it('Can successfully parse records', () => {
+      const zodAst = secureEvaluateSchema(`z.record(z.string(), z.number())`)
+      const mapper = new ZodToTypescriptMapper({fieldName: 'someKey'})
+      const rendered = mapper.renderField(zodAst.schema!)
+      expect(rendered).to.equal('"someKey": Record<string, number>')
+    })
+
+    it('Can successfully parse records with object values', () => {
+      const zodAst = secureEvaluateSchema(`z.record(z.string(), z.object({ id: z.string(), count: z.number() }))`)
+      const mapper = new ZodToTypescriptMapper({fieldName: 'someKey'})
+      const rendered = mapper.renderField(zodAst.schema!)
+      expect(rendered).to.equal('"someKey": Record<string, { "id": string; "count": number }>')
+    })
+
     it('Can successfully parse an optional wrapped type', () => {
       const zodAst = secureEvaluateSchema(`z.optional(z.string())`)