+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Added support for `ZodRecord` types in code generation mappers
+- Added `-o` / `--output-directory` flag to `generate` command to specify output directory per run
+
 ## 0.0.7 - 2025-10-20
 
 - Adds "whoami" command and optional verbose logging to debug login process [#65]